Improved tests structure Added tests coverage
