+++ tests/conftest.py
@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
 import pytest
+import asyncio
 
 from sanic import Sanic
+from wtforms.validators import ValidationError
 
+
+SLEEP_DURATION = 0.01
 
 @pytest.fixture(scope='function')
 def app():
@@ -15,3 +19,20 @@ def app():
         request[name] = session
 
     return test_app
+
+@pytest.fixture(scope='function')
+def async_validator_conditionally_fail():
+    async def async_validator(form, field):
+        await asyncio.sleep(SLEEP_DURATION)
+        if form.msg.data == 'fail':
+            raise ValidationError('Invalid')
+        else:
+            pass
+
+    return async_validator
+
+@pytest.fixture(scope='function')
+def async_validator_always_pass():
+    async def async_validator(form, field):
+        await asyncio.sleep(SLEEP_DURATION)
+    return async_validator
