+++ tests/test_form.py
@@ -8,21 +8,7 @@ from wtforms.validators import DataRequired, Length, ValidationError
 from wtforms import FileField, StringField, SubmitField
 
 from sanic_wtf import SanicForm, to_bytes
-
-
-# NOTE
-# taking shortcut here, assuming there will be only one "string" (the token)
-# ever longer than 40.
-csrf_token_pattern = '''value="([0-9a-f#]{40,})"'''
-
-
-def render_form(form, multipart=False):
-    if multipart is True:
-        multipart = ' enctype="multipart/form-data"'
-    return """
-    <form action="" method="POST"{}>
-    {}
-    </form>""".format(multipart, ''.join(str(field) for field in form))
+from .helpers import render_form, csrf_token_pattern
 
 
 def test_form_validation(app):
@@ -173,163 +159,6 @@ def test_validate_on_submit(app):
     req, resp = app.test_client.post('/', data=payload)
     assert resp.status == 200
     assert 'validated' in resp.text
-
-def test_validate_on_submit_async(app):
-    app.config['WTF_CSRF_SECRET_KEY'] = 'top secret !!!'
-
-    async def async_validator(form, field):
-        await asyncio.sleep(1)
-        if form.msg.data == 'fail':
-            raise ValidationError('Should fail')
-        else:
-            pass
-
-    class TestForm(SanicForm):
-        msg = StringField('Note', validators=[DataRequired(), Length(max=10), async_validator])
-        submit = SubmitField('Submit')
-
-    @app.route('/', methods=['POST'])
-    async def index(request):
-        form = TestForm(request)
-        if not await form.validate_on_submit_async():
-            return response.text('invalid')
-        else:
-            return response.text('valid')
-
-    @app.route('/', methods=['GET'])
-    async def index_(request):
-        form = TestForm(request)
-        content = render_form(form)
-        return response.html(content)
-
-    req, resp = app.test_client.get('/')
-    assert resp.status == 200
-    assert 'csrf_token' in resp.text
-    token = re.findall(csrf_token_pattern, resp.text)[0]
-    assert token
-
-    payload = {'msg': 'happy', 'csrf_token': token}
-    req, resp = app.test_client.post('/', data=payload)
-    assert resp.status == 200
-    assert 'valid' in resp.text
-
-def test_validate_on_submit_async_fails(app):
-    app.config['WTF_CSRF_SECRET_KEY'] = 'top secret !!!'
-
-    async def async_validator(form, field):
-        await asyncio.sleep(1)
-        if form.msg.data == 'fail':
-            raise ValidationError('Should fail')
-        else:
-            pass
-
-    async def async_validator_pass(form, field):
-        await asyncio.sleep(1)
-
-    class TestForm(SanicForm):
-        msg = StringField('Note', validators=[DataRequired(), Length(max=10), async_validator, async_validator_pass])
-        submit = SubmitField('Submit')
-
-    @app.route('/', methods=['POST'])
-    async def index(request):
-        form = TestForm(request)
-        if not await form.validate_on_submit_async():
-            return response.text('invalid')
-        else:
-            return response.text('valid')
-
-    @app.route('/', methods=['GET'])
-    async def index_(request):
-        form = TestForm(request)
-        content = render_form(form)
-        return response.html(content)
-
-    req, resp = app.test_client.get('/')
-    assert resp.status == 200
-    assert 'csrf_token' in resp.text
-    token = re.findall(csrf_token_pattern, resp.text)[0]
-    assert token
-
-    payload = {'msg': 'fail', 'csrf_token': token}
-    req, resp = app.test_client.post('/', data=payload)
-    assert resp.status == 200
-    assert 'invalid' in resp.text
-
-def test_validate_on_submit_async_and_sync(app):
-    app.config['WTF_CSRF_SECRET_KEY'] = 'top secret !!!'
-
-    async def async_validator(form, field):
-        await asyncio.sleep(1)
-
-    def sync_validator(form, field):
-        raise ValidationError('Should fail')
-
-    class TestForm(SanicForm):
-        msg = StringField('Note', validators=[DataRequired(), Length(max=10), async_validator, sync_validator])
-        submit = SubmitField('Submit')
-
-    @app.route('/', methods=['POST'])
-    async def index(request):
-        form = TestForm(request)
-        if not await form.validate_on_submit_async():
-            return response.text('invalid')
-        else:
-            return response.text('valid')
-
-    @app.route('/', methods=['GET'])
-    async def index_(request):
-        form = TestForm(request)
-        content = render_form(form)
-        return response.html(content)
-
-    req, resp = app.test_client.get('/')
-    assert resp.status == 200
-    assert 'csrf_token' in resp.text
-    token = re.findall(csrf_token_pattern, resp.text)[0]
-    assert token
-
-    payload = {'msg': 'fail', 'csrf_token': token}
-    req, resp = app.test_client.post('/', data=payload)
-    assert resp.status == 200
-    assert 'invalid' in resp.text
-
-def test_validate_on_submit_async_and_sync_stock_validator(app):
-    app.config['WTF_CSRF_SECRET_KEY'] = 'top secret !!!'
-
-    async def async_validator(form, field):
-        await asyncio.sleep(1)
-
-    def sync_validator(form, field):
-        raise ValidationError('Should fail')
-
-    class TestForm(SanicForm):
-        msg = StringField('Note', validators=[DataRequired(), Length(max=2), async_validator])
-        submit = SubmitField('Submit')
-
-    @app.route('/', methods=['POST'])
-    async def index(request):
-        form = TestForm(request)
-        if not await form.validate_on_submit_async():
-            return response.text('invalid')
-        else:
-            return response.text('valid')
-
-    @app.route('/', methods=['GET'])
-    async def index_(request):
-        form = TestForm(request)
-        content = render_form(form)
-        return response.html(content)
-
-    req, resp = app.test_client.get('/')
-    assert resp.status == 200
-    assert 'csrf_token' in resp.text
-    token = re.findall(csrf_token_pattern, resp.text)[0]
-    assert token
-
-    payload = {'msg': 'fail', 'csrf_token': token}
-    req, resp = app.test_client.post('/', data=payload)
-    assert resp.status == 200
-    assert 'invalid' in resp.text
 
 def test_file_upload(app):
     app.config['WTF_CSRF_ENABLED'] = False
