+++ tox.ini
@@ -3,4 +3,4 @@ envlist = py35, py36, py37
 
 [testenv]
 deps = -rrequirements-dev.txt
-commands = py.test
+commands = py.test -v --cov sanic_wtf -p no:warnings  # Disable aiohttp's verify_ssl warning
